@@ -1,27 +1,33 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+    xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
     <parent>
-        <groupId>org.springframework.boot</groupId>
-        <artifactId>spring-boot-starter-parent</artifactId>
-        <version>2.3.1.RELEASE</version>
-        <relativePath/> <!-- lookup parent from repository -->
+        <groupId>com.baeldung.docker</groupId>
+        <artifactId>docker</artifactId>
+        <version>0.0.1</version>
     </parent>
-    <groupId>com.baeldung.docker</groupId>
-    <artifactId>spring-boot-docker</artifactId>
-    <version>0.0.1-SNAPSHOT</version>
-    <name>spring-boot-docker</name>
+
+    <artifactId>docker-spring-boot</artifactId>
+
+    <name>docker-spring-boot</name>
     <description>Demo project showing Spring Boot and Docker</description>
 
     <properties>
-        <java.version>8</java.version>
+        <java.version>11</java.version>
     </properties>
 
     <dependencies>
         <dependency>
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-starter-web</artifactId>
+        </dependency>
+
+        <dependency>
+            <groupId>com.baeldung.docker</groupId>
+            <artifactId>docker-internal-dto</artifactId>
+            <version>0.0.1</version>
         </dependency>
 
         <dependency>
@@ -45,6 +51,7 @@
                 <configuration>
                     <layers>
                         <enabled>true</enabled>
+                        <configuration>${project.basedir}/src/layers.xml</configuration>
                     </layers>
                 </configuration>
             </plugin>
